Refactor OpenStreetMap::search to return an array of JsonSearchResults
After using the method in a small project that I've been building, it
didn't make sense for that project to have to know so much about the
internals of a result and to have to work with them so manually. Given
that, this change refactors the method, and the related code, to return
both the raw result in the body of the response from the API and to also
return a list of JsonSearchResult objects.

At the moment, that type is fixed, as JSON is the only format that's
properly handled. However, that will change as more of the other
available formats are supported.