Add a class to simplify managing search options
I don't know if this is the best approach to take, but the more I've
been reviewing the documentation for the Search endpoint, the more
options I see, which left me wondering what the best way was for them to
be provided. So far, this is the best solution that I've come up with,
though it may change in the future.